requirements.txt
azure-ai-evaluation
azure-identity
azure-ai-projects
azure-ai-ml
pandas
pytest